Display a snarky message on divide by zero attempt
Before, number / zero resulted in Infinity and zero / zero resulted in
NaN, both of which could be operated on further due to storage in
operandOne. This is undesirable. Now the message 'you nutter' is
displayed and cannot be operated on as all data is cleared.

Also, enable strict mode in the script for improved code quality.