Fix text wrap of expression with minus
The issue was with displaying an expression with minus in it. If the
expression overflowed the display (the p element), the second operand
was moved below the first operand and the operator, essentially
wrapping it. This was not noted with other operators.

The solution was to set a declaration in the p ruleset such that the
text in it would not wrap. The property used was relatively new.

The exact cause of the wrapping was not found.